Modify adaptador de lista para mostrar las tasks desactivadas y la task siguiente

+++ FSW-Huella/app/src/main/java/com/fime/fsw/huella/huella/Fragments/RecorridoActualFragment.java
@@ -94,7 +94,7 @@ public class RecorridoActualFragment extends Fragment {
         rvRecorrido.setHasFixedSize(true);
         rvRecorrido.setLayoutManager(linearLayoutManager);
 
-        int currentItem = mSesion.getCurrentItemLista();
+        long currentItem = mSesion.getCurrentItemLista();
 
         //Se obtiene la info de nuestro Realm
         final OrderedRealmCollection<Task> recorridoData = getAllRealmTasks();
